lesson1_streamlit_llm/requirements.txt
streamlit
python-dotenv
groq
openai